refactor(src/app/api/backend/updateExerciseStatus/route.ts src/components/PageElements/Renders/RenderDetailingTopicPage/index.tsx src/components/StatusSelect/index.tsx  src/components/fetchStatus/fecthStatusTopic/index.ts src/components/fetchStatus/index.ts -> src/components/fetchStatus/fetchStatusExercise/index.ts) corrigimos o update status e dividimos o fecht status em duas pastas, uma para exercicio e outra pára topico @Lip29bm @stephanybrazeir0 @Geoziihdev
