+++ src/components/fetchStatus/fetchStatusExercise/index.ts
@@ -1,19 +1,16 @@
-import { useState, useEffect, useCallback, useRef } from "react";
+import { useState, useEffect, useCallback} from "react";
 
 interface UseStatusProps {
   topicId: string;
   itemId: string;
-  pageId?: string;
 }
 
 export const useStatus = ({
   topicId,
   itemId,
-  pageId,
 }: UseStatusProps) => {
   const [status, setStatus] = useState<string>("NotStarted");
   const [isLoading, setIsLoading] = useState<boolean>(false);
-  const topicIdsRef = useRef<string[] | null>(null);
 
   const fetchStatus = useCallback(async () => {
     if (!topicId || !itemId) return;
@@ -46,18 +43,15 @@ export const useStatus = ({
   const updateStatus = async (newStatus: string) => {
     setIsLoading(true);
     setStatus(newStatus);
-
     try {
       const response = await fetch("/api/backend/updateExerciseStatus", {
         method: "PUT",
         headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({
           topicId,
           itemId,
           itemStatus: newStatus,
-        }),
+          "Content-Type": "application/json",
+        },
       });
 
       if (!response.ok) throw new Error(`Erro ${response.status}`);
@@ -67,44 +61,10 @@ export const useStatus = ({
       setIsLoading(false);
     }
   };
-
-  const getTopicStatus = useCallback(async () => {
-    if (!topicId || !pageId) return;
-    
-    if (JSON.stringify([topicId]) === JSON.stringify(topicIdsRef.current)) {
-      return;
-    }
-
-    topicIdsRef.current = [topicId];
-
-    try {
-      const response = await fetch(`/api/backend/getTopicExercisesStatus`, {
-        method: "GET",
-        headers: {
-          "Content-Type": "application/json",
-          topicId,
-        },
-      });
-
-      if (!response.ok) throw new Error(`Erro ${response.status}`)
-
-      const data = await response.json();
-    
-      data.status.map((element: any) =>{
-          if (element.itemId === pageId){
-            setStatus(element.itemStatus)
-          }
-      });
-         
-    } catch (error) {
-      console.error("Erro ao buscar status dos exercícios do tópico:", error);
-    }
-  }, [topicId, pageId]);
-
+  
   useEffect(() => {
     fetchStatus();
-    getTopicStatus();
-  }, [fetchStatus, getTopicStatus]);
+  }, [fetchStatus]);
 
   return { status, isLoading, updateStatus };
 };
